点選択時の初期化処理を書き出し
